fonctions serveurs deplacées pour gagner en lisibilité
